phase4-kuber/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cohere==4.9.0
python-multipart==0.0.6
alembic==1.13.1
asyncpg==0.29.0
python-dotenv==1.0.0
requests==2.31.0